+++ HTMLReader/Code/HTMLSelector.m
@@ -24,7 +24,7 @@ static NSError * ParseError(NSString *reason, NSString *string, NSUInteger posit
      */
     NSString *caretString = [@"^" stringByPaddingToLength:position+1 withString:@" " startingAtIndex:0];
     NSString *failureReason = [NSString stringWithFormat:@"Error near character %zd: %@\n\n\t%@\n\t\%@",
-                               position,
+                               ( long ) position,
                                reason,
                                string,
                                caretString];
